+++ httprequest_lego_provider/tests/__init__ .py
@@ -1,3 +1,3 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
-"""Unit tests for the lego module."""
+"""Unit tests for the HTTPRequest Lego Provider module."""
