+++ tox.ini
@@ -7,8 +7,8 @@ skip_missing_interpreters = True
 envlist = lint, unit, static, coverage-report
 
 [vars]
-src_path = {toxinidir}/lego/
-tst_path = {toxinidir}/lego/tests/
+src_path = {toxinidir}/httprequest_lego_provider/
+tst_path = {toxinidir}/httprequest_lego_provider/tests/
 ;lib_path = {toxinidir}/lib/charms/operator_name_with_underscores
 all_path = {[vars]src_path}
